pyproject.toml
[tool.poetry]
name = "imoveis-api"
version = "0.1.0"
description = ""
authors = ["Matheus Vargas <matheusvargas042@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
pydantic-settings = "^2.7.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.10.1"
phonenumbers = "^8.13.55"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.3"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
testcontainers = "^4.9.1"
pytest-asyncio = "^0.25.3"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev imoveis_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=imoveis_api -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
